Clean up RPM build definitions
The opendaylight.spec template had changed (to add clean uninstalls)
since these build_vars sections were added, so the generated specs used
by the released RPMs differ from the ones that would currently be
have been supplanted by more recent ones, just remove them vs updating,
rebuilding and re-releasing.

+++ rpm/specs/opendaylight-4.0.0-1.rc1.spec
@@ -2,19 +2,19 @@
 %define __jar_repack 0
 
 Name:       opendaylight
-Version:    3.0.0
+Version:    4.0.0
 # The Fedora/CentOS packaging guidelines *require* the use of a disttag. ODL's
 #   RPM build doesn't do anything Fedora/CentOS specific, so the disttag is
 #   unnecessary and unused in our case, but both the docs and the pros (apevec)
 #   agree that we should include it.
 # See: https://fedoraproject.org/wiki/Packaging:DistTag
-Release:    2.el7
+Release:    1.rc1.el7
 BuildArch:  noarch
 Summary:    OpenDaylight SDN Controller
 Group:      Applications/Communications
 License:    EPL-1.0
 URL:        http://www.opendaylight.org
-Source0:    https://nexus.opendaylight.org/content/groups/public/org/opendaylight/integration/distribution-karaf/0.3.0-Lithium/distribution-karaf-0.3.0-Lithium.tar.gz
+Source0:    https://nexus.opendaylight.org/content/groups/public/org/opendaylight/integration/distribution-karaf/0.4.0-Beryllium-RC1/distribution-karaf-0.4.0-Beryllium-RC1.tar.gz
 Source1:    %name-c6d7ee1f21d62ff8ffb741c7a12c3d8d3aa3d6ed.service.tar.gz
 Buildroot:  /tmp
 # Required for ODL at run time
@@ -32,11 +32,11 @@ getent passwd odl > /dev/null || useradd odl -M -d $RPM_BUILD_ROOT/opt/%name
 getent group odl > /dev/null || groupadd odl
 
 %description
-OpenDaylight Lithium
+OpenDaylight Beryllium-RC1
 
 %prep
 # Extract Source0 (ODL archive)
-%autosetup -n distribution-karaf-0.3.0-Lithium
+%autosetup -n distribution-karaf-0.4.0-Beryllium-RC1
 # Extract Source1 (systemd config)
 %autosetup -T -D -b 1 -c -n %name-c6d7ee1f21d62ff8ffb741c7a12c3d8d3aa3d6ed.service
 
@@ -44,7 +44,7 @@ OpenDaylight Lithium
 # Create directory in build root for ODL
 mkdir -p $RPM_BUILD_ROOT/opt/%name
 # Copy ODL from archive to its dir in build root
-cp -r ../distribution-karaf-0.3.0-Lithium/* $RPM_BUILD_ROOT/opt/%name
+cp -r ../distribution-karaf-0.4.0-Beryllium-RC1/* $RPM_BUILD_ROOT/opt/%name
 # Create directory in build root for systemd .service file
 mkdir -p $RPM_BUILD_ROOT/%{_unitdir}
 # Copy ODL's systemd .service file to correct dir in build root
@@ -54,7 +54,9 @@ cp ../../BUILD/%name-c6d7ee1f21d62ff8ffb741c7a12c3d8d3aa3d6ed.service/%name-c6d7
 # When the RPM is removed, the subdirs containing new files wouldn't normally
 #   be deleted. Manually clean them up.
 #   Warning: This does assume there's no data there that should be preserved
-rm -rf $RPM_BUILD_ROOT/opt/%name
+if [ $1 -eq 0 ]; then
+    rm -rf $RPM_BUILD_ROOT/opt/%name
+fi
 
 %files
 # ODL will run as odl:odl, set as user:group for ODL dir, don't override mode
@@ -64,5 +66,5 @@ rm -rf $RPM_BUILD_ROOT/opt/%name
 
 
 %changelog
-* Fri Jul 17 2015 Daniel Farrell <dfarrell@redhat.com> - 3.0.0-2
-- Include required disttag in RPM release version
+* Wed Jan 22 2016 Daniel Farrell <dfarrell@redhat.com> - 4.0.0-1.rc1
+- Create Beryllium RC1 RPM
